[cmd/telemetrygen] move validate func out of common
Removing common package. Moving the code to validate package instead.
